+++ src/users.ts
@@ -3,11 +3,11 @@ import AWS from 'aws-sdk';
 const dynamoDB = new AWS.DynamoDB.DocumentClient();
 const installationsKeyValueStore = 'Boost.GitHub-App.installations';
 
-export async function getUser(email: string): Promise<{ installationId: string; username: string } | undefined> {
+export async function getUser(account: string): Promise<{ installationId: string; username: string } | undefined> {
     try {
         const params = {
             TableName: installationsKeyValueStore,
-            Key: { email } // primary key
+            Key: { account } // primary key
         };
 
         const userInfo = await dynamoDB.get(params).promise();
@@ -24,11 +24,11 @@ export async function getUser(email: string): Promise<{ installationId: string; 
     }
 }
 
-export async function saveUser(email: string, installationId: string, username: string): Promise<void> {
+export async function saveUser(account: string, installationId: string, username: string): Promise<void> {
     const params = {
         TableName: installationsKeyValueStore,
         Item: {
-            email, // primary key
+            account, // primary key
             installationId,
             username,
         },